Preliminary HTTPS support
Instead of tunneling the socket provided on the `connect` event to the
target server, connect to a local HTTPS server and present a self-signed
certificate (ideally one that matches the target hostname and is signed
with a CA that's been installed--check mitmproxy to see how this is done
dynamically) before proxying the request normally.

This will allow the `ssl` option to be deprecated and for proxy clients
to make HTTPS connections directly and have them signed.